backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
python-decouple>=3.8
requests>=2.28.0
dj-database-url>=2.1.0
gunicorn>=21.2.0
